% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_sample_means.R
\name{many_sample_means}
\alias{many_sample_means}
\title{Performs the sample_mean process many times (reps) and returns a vector of the results}
\usage{
many_sample_means(vec, n, reps)
}
\arguments{
\item{vec}{A vector}

\item{n}{An integer}

\item{reps}{An integer}
}
\value{
A vector of the results of sample mean being called reps times
}
\description{
Performs the sample_mean process many times (reps) and returns a vector of the results
}
