% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Returns mean of random sample of size n}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{vec}{a numeric vector}

\item{n}{A integer}
}
\value{
Mean of random sample of size n of vec
}
\description{
Returns mean of random sample of size n
}
